moved private helpers to helper file
